Update glibc on AT next and fix FVTR test
Bump to revision e766ce3088e1

Starting with glibc commit '8208be389b * Install shared objects under their ABI
names', glibc does not install libc-<version>.so files anymore, instead
libc.so.6 became a regular file (was a symlink before).

This breaks systemtap.exp because it looks for the old-style .so name. This
commit updates the test to make sure it uses the correct .so name for glibc >=
2.34.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>