Disable use of morecore on libhugetlbfs if not supported by glibc
Starting with glibc 2.32, __morecore hook has been marked as deprecated, and was
completely removed on glibc 2.34, which causes an undefined symbol error during
the build of libhugetlbfs. This affects AT 15 and current AT next.

Greater changes are needed in order to keep providing the same functionality
with future versions of glibc. Meanwhile, we can disable hugepage-backed heap
setup if __morecore is not available so users can at least keep using the other
features provided by the library.

In addition to disabling morecore, we also apply a patch to force make to always
use bash as the shell when building libhugetlbfs. By default make uses /bin/sh,
which is a symlink to dash on Debian and Ubuntu, and that has issues when
parsing the Makefile.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>