Disable debuginfod for binutils build
If the libraries for `debuginfod` are installed on the build system,
binutils will enable support for it, which brings in a raft of other
dependencies.  This is normal, but one of the dependencies causes
problems in a specific situation.  On Red Hat Enterprise Linux 8,

The `readelf` built for `binutils` will depend on `libdebuginfod`
which in turn depends on `libcurl` which in turn depends on
`libcrypto` and `libk5crypto`.  The Red Hat Enterprise Linux (RHEL) 8
version of `libk5crypto` depends on symbols in `libcrypto` which are
out-of-tree patches, and thus do not appear in the AT-built version
of `libcrypto`. The AT-build `readelf` is not usable as a result:
```
$ $AT_DEST/bin/readelf
[AT_DEST]/bin/readelf: symbol lookup error: /lib64/libk5crypto.so.3: undefined symbol: EVP_KDF_ctrl, version OPENSSL_1_1_1b
```

One option is to backport those same out-of-tree patches to AT,
but that adds a support burden in a security-sensitive area.

For now, disable the use of `debuginfod` in the AT-built binutils
for RHEL 8.

Fixes #1969.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>