Disable some Python FVTR tests that fail intermittently
Some Python tests have been failing intermittently for a while (#1873 and #2368),
most of (all?) the times on a single RHEL8 host we use for automated testing. We
have not been able to pinpoint the issue so far. But it could be related to high
system load as these tests are time-sensitive and are failing because of expired
timeouts. In most cases a simple rebuild makes the build pass.

This commit adds the following tests to the list of excluded tests:
     - test_eintr
     - test_ftplib
     - test_ssl
     - test_subprocess

Closes #1873
Closes #2368

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>