Build and package optimized libpython.so
The Python builds are currently optimized only for the default
(earliest supported) architecture level for the GCC in use.

Optimizations which are _not_ performed are compile-time optimizations
for the more recent / latest processor architectures. For Power, there
are significant optimizations possible by using `-mcpu=power9` or
`-mcpu=power10`.

Conveniently, the (`--with-shared`) `python` binary is linked with its
`libpython.so` such that the normal library search is done to find it.
This search process includes the capability to add processor-optimized
versions of the library. For example, if such files are installed thus:
- `/opt/atXX.0/lib64/libpython.so`
- `/opt/atXX.0/lib64/glibc-hwcaps/power9/libpython.so`
- `/opt/atXX.0/lib64/glibc-hwcaps/power10/libpython.so`

On a Power8 system, the `lib64/libpython.so` will be used.
On a Power9 system, the `lib64/glibc-hwcaps/power9/libpython.so`
will be used.
On a Power10 system, the `lib64/glibc-hwcaps/power10/libpython.so`
will be used.
On a theoretical Power[n+1] system, the latest
`lib64/glibc-hwcaps/power[n]/libpython.so` will be used.

The same is not true for Python modules. These are dynamically loaded
using a fixed path name and do not use a normal library search.
Python modules are not covered by these changes.

Perform additional builds of Python `libpython3*.so*` with
processor-specific optimization and copy the built objects to the
appropriate place in the build tree to be incorporated in the final AT
packages.

The python objects are currently built with:
```
--enable-optimizations
```

The impact of this flag is to run the built binaries under some
predetermined benchmarks and optimize the binaries based on the
measurements collected during those runs. This is also known as
profile-directed or profile-guided optimization (PGO).

Note that any PGO optimizations are obviously based on the generated
binaries, and will be specific to the processor on which the
build is performed.

Since the build process runs the built binaries, it is not possible
to build PGO-optimized binaries for Power10 on a Power9 or Power8
system, or for Power9 on a Power8 system, so remove the
`--enable-optimizations` flag so the builds can be performed anywhere,
with the loss of PGO improvements.

`stage_optimized` differs from `stage_1`:

- The "configure" function has the optimization flags in a new `OPT`
  variable and adding `-mcpu=[processor]` flag for processor-specific
  optimization.

- The "install" function copies the optimized `libpython.so` to the
  appropriate place in the AT destination directory tree, under
  `lib64/glibc-hwcaps/[processor]` so they can be automatically
  loaded at run-time.

- No `atcfg_make_check`.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>