next: Ignore -Waddress on binutils_2 build
GCC started to warn about an unnecessary comparison on readelf.c:find_section().

binutils/readelf.c: In function ‘find_section’:
binutils/readelf.c:349:8: error: the comparison will always evaluate as ‘true’
for the pointer operand in ‘filedata->section_headers + (sizetype)((long
unsigned int)i * 80)’ must not be NULL [-Werror=address]
  349 |   ((X) != NULL                                                          \
      |

The function first checks if filedata->section_headers is NULL, and then starts
iterating on each element in the array. Inside the loop, each element is also
checked for NULL as part of a macro, which ends up being a redundant check for
the first element.

We can safely ignore that for now until this is fixed upstream.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>