Update gcc and glibc on AT next + fixes
gcc:
Bump to version 12.0.1
Bump to revision fac15bf84807

glibc:
Bump to version 2.35.9000
Bump to revision 098c795e85fb

We need to bump the two at the same time because of a new warning
introduced by GCC, which is triggering a build failured on glibc. The
fix is already on glibc, but it depends on the warning existing on
GCC.

In the set of changes included by this GCC bump, there's commit
'3f30f2d1dbb rs6000: Fix LE code gen for vec_cnt[lt]z_lsbb [PR95082]',
which breaks one of the zlib optimizations distributed on AT. So I had
to update the zlib optimizations patch with a fix for it.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>