Set AT_BASE before running FVTR
This patch set AT_BASE using the path of the running fvtr.sh script.
AT_BASE is no longer stored in the config_at* file.

This avoids using AT_BASE whereas it points to a directory that dosen't exist,
which could happen when FVTR is run against installed AT packages, with a
different working directory than the one used for the build.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>