Update installation guidelines
Added missing installation guidelines for Debian 11, Ubuntu 22.04 and RHEL9
in the release notes.

Fix #3055

Also, replaced most of the GPG keys with a `xxxxxxxx` pattern. This pattern
is replaced with the appropriate GPG key value when generating the final
release notes.

In addition, rearranged the sections ; a fedora section was in middle of
ubuntu sections.

-
Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>